pyproject.toml
[project]
name = "claimsense-ai"
version = "0.1.0"
description = "AI-powered claim denial intelligence system built with FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claimsense"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

